+++ site/src/pages/authors.js
@@ -4,7 +4,7 @@ import Layout from "../components/layout";
 import SeoComponent from "../components/seo";
 import Navbar from "../components/Navbar";
 import Footer from "../components/Footer";
-import AboutL5io from "../sections/aboutl5io";
+import AboutL5io from "../sections/AboutLayer5/index";
 
 const Authors = () => {
   return (
